+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
   * **Fixed:**
     * Crash when crafting vanilla recipe with recipe function in Artisan's table [(#240)](https://github.com/codetaylor/artisan-worktables/issues/240)
+    * Improper handling of remaining items [(#241)](https://github.com/codetaylor/artisan-worktables/issues/241)
     
 ---